Adiciona generador de identificadores y cambia identificador al tipo long

+++ Bja.Entidades/Permission.cs
@@ -8,8 +8,8 @@ namespace Bja.Entidades
 {
     public class Permission
     {
-        public Int32 Id { get; set; }
-        public Int32 IdInstance { get; set; }
+        public long Id { get; set; }
+        public long IdInstance { get; set; }
         public String Name { get; set; }
         public String Description { get; set; }
